+++ src/main/java/edu/dhu/mooc/MoocApplication.java
@@ -1,5 +1,6 @@
 package edu.dhu.mooc;
 
+import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
