--- src/main/java/edu/dhu/mooc/entity/TeacherExample.java
@@ -1,722 +1,0 @@
-package edu.dhu.mooc.entity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class TeacherExample {
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    protected String orderByClause;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    protected boolean distinct;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    protected List<Criteria> oredCriteria;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public TeacherExample() {
-        oredCriteria = new ArrayList<Criteria>();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public void setOrderByClause(String orderByClause) {
-        this.orderByClause = orderByClause;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public String getOrderByClause() {
-        return orderByClause;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public void setDistinct(boolean distinct) {
-        this.distinct = distinct;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public boolean isDistinct() {
-        return distinct;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public List<Criteria> getOredCriteria() {
-        return oredCriteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public void or(Criteria criteria) {
-        oredCriteria.add(criteria);
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public Criteria or() {
-        Criteria criteria = createCriteriaInternal();
-        oredCriteria.add(criteria);
-        return criteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public Criteria createCriteria() {
-        Criteria criteria = createCriteriaInternal();
-        if (oredCriteria.size() == 0) {
-            oredCriteria.add(criteria);
-        }
-        return criteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    protected Criteria createCriteriaInternal() {
-        Criteria criteria = new Criteria();
-        return criteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public void clear() {
-        oredCriteria.clear();
-        orderByClause = null;
-        distinct = false;
-    }
-
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    protected abstract static class GeneratedCriteria {
-        protected List<Criterion> criteria;
-
-        protected GeneratedCriteria() {
-            super();
-            criteria = new ArrayList<Criterion>();
-        }
-
-        public boolean isValid() {
-            return criteria.size() > 0;
-        }
-
-        public List<Criterion> getAllCriteria() {
-            return criteria;
-        }
-
-        public List<Criterion> getCriteria() {
-            return criteria;
-        }
-
-        protected void addCriterion(String condition) {
-            if (condition == null) {
-                throw new RuntimeException("Value for condition cannot be null");
-            }
-            criteria.add(new Criterion(condition));
-        }
-
-        protected void addCriterion(String condition, Object value, String property) {
-            if (value == null) {
-                throw new RuntimeException("Value for " + property + " cannot be null");
-            }
-            criteria.add(new Criterion(condition, value));
-        }
-
-        protected void addCriterion(String condition, Object value1, Object value2, String property) {
-            if (value1 == null || value2 == null) {
-                throw new RuntimeException("Between values for " + property + " cannot be null");
-            }
-            criteria.add(new Criterion(condition, value1, value2));
-        }
-
-        public Criteria andTIdIsNull() {
-            addCriterion("t_id is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdIsNotNull() {
-            addCriterion("t_id is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdEqualTo(String value) {
-            addCriterion("t_id =", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdNotEqualTo(String value) {
-            addCriterion("t_id <>", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdGreaterThan(String value) {
-            addCriterion("t_id >", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdGreaterThanOrEqualTo(String value) {
-            addCriterion("t_id >=", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdLessThan(String value) {
-            addCriterion("t_id <", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdLessThanOrEqualTo(String value) {
-            addCriterion("t_id <=", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdLike(String value) {
-            addCriterion("t_id like", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdNotLike(String value) {
-            addCriterion("t_id not like", value, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdIn(List<String> values) {
-            addCriterion("t_id in", values, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdNotIn(List<String> values) {
-            addCriterion("t_id not in", values, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdBetween(String value1, String value2) {
-            addCriterion("t_id between", value1, value2, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTIdNotBetween(String value1, String value2) {
-            addCriterion("t_id not between", value1, value2, "tId");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameIsNull() {
-            addCriterion("t_name is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameIsNotNull() {
-            addCriterion("t_name is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameEqualTo(String value) {
-            addCriterion("t_name =", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameNotEqualTo(String value) {
-            addCriterion("t_name <>", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameGreaterThan(String value) {
-            addCriterion("t_name >", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameGreaterThanOrEqualTo(String value) {
-            addCriterion("t_name >=", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameLessThan(String value) {
-            addCriterion("t_name <", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameLessThanOrEqualTo(String value) {
-            addCriterion("t_name <=", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameLike(String value) {
-            addCriterion("t_name like", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameNotLike(String value) {
-            addCriterion("t_name not like", value, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameIn(List<String> values) {
-            addCriterion("t_name in", values, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameNotIn(List<String> values) {
-            addCriterion("t_name not in", values, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameBetween(String value1, String value2) {
-            addCriterion("t_name between", value1, value2, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNameNotBetween(String value1, String value2) {
-            addCriterion("t_name not between", value1, value2, "tName");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordIsNull() {
-            addCriterion("password is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordIsNotNull() {
-            addCriterion("password is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordEqualTo(String value) {
-            addCriterion("password =", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordNotEqualTo(String value) {
-            addCriterion("password <>", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordGreaterThan(String value) {
-            addCriterion("password >", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordGreaterThanOrEqualTo(String value) {
-            addCriterion("password >=", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordLessThan(String value) {
-            addCriterion("password <", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordLessThanOrEqualTo(String value) {
-            addCriterion("password <=", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordLike(String value) {
-            addCriterion("password like", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordNotLike(String value) {
-            addCriterion("password not like", value, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordIn(List<String> values) {
-            addCriterion("password in", values, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordNotIn(List<String> values) {
-            addCriterion("password not in", values, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordBetween(String value1, String value2) {
-            addCriterion("password between", value1, value2, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andPasswordNotBetween(String value1, String value2) {
-            addCriterion("password not between", value1, value2, "password");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderIsNull() {
-            addCriterion("gender is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderIsNotNull() {
-            addCriterion("gender is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderEqualTo(String value) {
-            addCriterion("gender =", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderNotEqualTo(String value) {
-            addCriterion("gender <>", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderGreaterThan(String value) {
-            addCriterion("gender >", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderGreaterThanOrEqualTo(String value) {
-            addCriterion("gender >=", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderLessThan(String value) {
-            addCriterion("gender <", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderLessThanOrEqualTo(String value) {
-            addCriterion("gender <=", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderLike(String value) {
-            addCriterion("gender like", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderNotLike(String value) {
-            addCriterion("gender not like", value, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderIn(List<String> values) {
-            addCriterion("gender in", values, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderNotIn(List<String> values) {
-            addCriterion("gender not in", values, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderBetween(String value1, String value2) {
-            addCriterion("gender between", value1, value2, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andGenderNotBetween(String value1, String value2) {
-            addCriterion("gender not between", value1, value2, "gender");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneIsNull() {
-            addCriterion("phone is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneIsNotNull() {
-            addCriterion("phone is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneEqualTo(String value) {
-            addCriterion("phone =", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneNotEqualTo(String value) {
-            addCriterion("phone <>", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneGreaterThan(String value) {
-            addCriterion("phone >", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneGreaterThanOrEqualTo(String value) {
-            addCriterion("phone >=", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneLessThan(String value) {
-            addCriterion("phone <", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneLessThanOrEqualTo(String value) {
-            addCriterion("phone <=", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneLike(String value) {
-            addCriterion("phone like", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneNotLike(String value) {
-            addCriterion("phone not like", value, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneIn(List<String> values) {
-            addCriterion("phone in", values, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneNotIn(List<String> values) {
-            addCriterion("phone not in", values, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneBetween(String value1, String value2) {
-            addCriterion("phone between", value1, value2, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andPhoneNotBetween(String value1, String value2) {
-            addCriterion("phone not between", value1, value2, "phone");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumIsNull() {
-            addCriterion("t_num is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumIsNotNull() {
-            addCriterion("t_num is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumEqualTo(String value) {
-            addCriterion("t_num =", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumNotEqualTo(String value) {
-            addCriterion("t_num <>", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumGreaterThan(String value) {
-            addCriterion("t_num >", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumGreaterThanOrEqualTo(String value) {
-            addCriterion("t_num >=", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumLessThan(String value) {
-            addCriterion("t_num <", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumLessThanOrEqualTo(String value) {
-            addCriterion("t_num <=", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumLike(String value) {
-            addCriterion("t_num like", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumNotLike(String value) {
-            addCriterion("t_num not like", value, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumIn(List<String> values) {
-            addCriterion("t_num in", values, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumNotIn(List<String> values) {
-            addCriterion("t_num not in", values, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumBetween(String value1, String value2) {
-            addCriterion("t_num between", value1, value2, "tNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andTNumNotBetween(String value1, String value2) {
-            addCriterion("t_num not between", value1, value2, "tNum");
-            return (Criteria) this;
-        }
-    }
-
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table teacher
-     *
-     * @mbg.generated do_not_delete_during_merge Sun Mar 14 16:13:17 CST 2021
-     */
-    public static class Criteria extends GeneratedCriteria {
-
-        protected Criteria() {
-            super();
-        }
-    }
-
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table teacher
-     *
-     * @mbg.generated Sun Mar 14 16:13:17 CST 2021
-     */
-    public static class Criterion {
-        private String condition;
-
-        private Object value;
-
-        private Object secondValue;
-
-        private boolean noValue;
-
-        private boolean singleValue;
-
-        private boolean betweenValue;
-
-        private boolean listValue;
-
-        private String typeHandler;
-
-        public String getCondition() {
-            return condition;
-        }
-
-        public Object getValue() {
-            return value;
-        }
-
-        public Object getSecondValue() {
-            return secondValue;
-        }
-
-        public boolean isNoValue() {
-            return noValue;
-        }
-
-        public boolean isSingleValue() {
-            return singleValue;
-        }
-
-        public boolean isBetweenValue() {
-            return betweenValue;
-        }
-
-        public boolean isListValue() {
-            return listValue;
-        }
-
-        public String getTypeHandler() {
-            return typeHandler;
-        }
-
-        protected Criterion(String condition) {
-            super();
-            this.condition = condition;
-            this.typeHandler = null;
-            this.noValue = true;
-        }
-
-        protected Criterion(String condition, Object value, String typeHandler) {
-            super();
-            this.condition = condition;
-            this.value = value;
-            this.typeHandler = typeHandler;
-            if (value instanceof List<?>) {
-                this.listValue = true;
-            } else {
-                this.singleValue = true;
-            }
-        }
-
-        protected Criterion(String condition, Object value) {
-            this(condition, value, null);
-        }
-
-        protected Criterion(String condition, Object value, Object secondValue, String typeHandler) {
-            super();
-            this.condition = condition;
-            this.value = value;
-            this.secondValue = secondValue;
-            this.typeHandler = typeHandler;
-            this.betweenValue = true;
-        }
-
-        protected Criterion(String condition, Object value, Object secondValue) {
-            this(condition, value, secondValue, null);
-        }
-    }
-}
