+++ src/main/resources/mappers/PartMapper.xml
@@ -1,0 +1,104 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="edu.dhu.mooc.dao.PartMapper">
+  <resultMap id="BaseResultMap" type="edu.dhu.mooc.entity.Part">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    <id column="p_id" jdbcType="VARCHAR" property="p_id" />
+    <result column="p_index" property="p_index"/>
+    <result column="p_name" jdbcType="VARCHAR" property="p_name" />
+  </resultMap>
+  <sql id="Base_Column_List">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    p_id, p_name
+  </sql>
+  <select id="selectByPrimaryKey" parameterType="java.lang.String" resultMap="BaseResultMap">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    select 
+    <include refid="Base_Column_List" />
+    from part
+    where p_id = #{p_id,jdbcType=VARCHAR}
+  </select>
+  <delete id="deleteByPrimaryKey" parameterType="java.lang.String">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    delete from part
+    where p_id = #{p_id,jdbcType=VARCHAR}
+  </delete>
+  <insert id="insert" parameterType="edu.dhu.mooc.entity.Part">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    insert into part (p_id, p_name)
+    values (#{p_id,jdbcType=VARCHAR}, #{p_name,jdbcType=VARCHAR})
+  </insert>
+  <insert id="insertSelective" parameterType="edu.dhu.mooc.entity.Part">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    insert into part
+    <trim prefix="(" suffix=")" suffixOverrides=",">
+      <if test="p_id != null">
+        p_id,
+      </if>
+      <if test="p_name != null">
+        p_name,
+      </if>
+    </trim>
+    <trim prefix="values (" suffix=")" suffixOverrides=",">
+      <if test="p_id != null">
+        #{p_id,jdbcType=VARCHAR},
+      </if>
+      <if test="p_name != null">
+        #{p_name,jdbcType=VARCHAR},
+      </if>
+    </trim>
+  </insert>
+  <update id="updateByPrimaryKeySelective" parameterType="edu.dhu.mooc.entity.Part">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    update part
+    <set>
+      <if test="p_name != null">
+        p_name = #{p_name,jdbcType=VARCHAR},
+      </if>
+    </set>
+    where p_id = #{p_id,jdbcType=VARCHAR}
+  </update>
+  <update id="updateByPrimaryKey" parameterType="edu.dhu.mooc.entity.Part">
+    <!--
+      WARNING - @mbg.generated
+      This element is automatically generated by MyBatis Generator, do not modify.
+      This element was generated on Mon Apr 12 15:50:40 CST 2021.
+    -->
+    update part
+    set p_name = #{p_name,jdbcType=VARCHAR}
+    where p_id = #{p_id,jdbcType=VARCHAR}
+  </update>
+
+  <select id="getPartListByCourseId" resultMap="BaseResultMap"
+          parameterType="string">
+    select * from part where c_id = #{c_id};
+  </select>
+</mapper>
