+++ src/main/resources/mappers/TeacherMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="edu.dhu.mooc.dao.TeacherMapper">
-  <resultMap id="BaseResultMap" type="edu.dhu.mooc.entity.Teacher">
+  <resultMap id="BaseResultMap" type="edu.dhu.mooc.entity.teacher.Teacher">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -44,7 +44,7 @@
     delete from teacher
     where t_id = #{t_id,jdbcType=VARCHAR}
   </delete>
-  <insert id="insert" parameterType="edu.dhu.mooc.entity.Teacher">
+  <insert id="insert" parameterType="edu.dhu.mooc.entity.teacher.Teacher">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -57,7 +57,7 @@
       #{gender,jdbcType=VARCHAR}, #{phone,jdbcType=VARCHAR}
       )
   </insert>
-  <insert id="insertSelective" parameterType="edu.dhu.mooc.entity.Teacher">
+  <insert id="insertSelective" parameterType="edu.dhu.mooc.entity.teacher.Teacher">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -99,7 +99,7 @@
       </if>
     </trim>
   </insert>
-  <update id="updateByPrimaryKeySelective" parameterType="edu.dhu.mooc.entity.Teacher">
+  <update id="updateByPrimaryKeySelective" parameterType="edu.dhu.mooc.entity.teacher.Teacher">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -122,7 +122,7 @@
     </set>
     where t_id = #{t_id,jdbcType=VARCHAR}
   </update>
-  <update id="updateByPrimaryKey" parameterType="edu.dhu.mooc.entity.Teacher">
+  <update id="updateByPrimaryKey" parameterType="edu.dhu.mooc.entity.teacher.Teacher">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -135,11 +135,49 @@
       phone = #{phone,jdbcType=VARCHAR},
     where t_id = #{t_id,jdbcType=VARCHAR}
   </update>
+
+
   <select id="selectByAccount" parameterType="java.lang.String" resultMap="BaseResultMap">
     select
     <include refid="Base_Column_List" />
     from teacher
     where phone = #{phone} and
     password = #{password}
+  </select>
+
+
+  <resultMap id="courseItemShownOnTeacherEnd" type="edu.dhu.mooc.entity.teacher.TeacherCourse">
+    <result column="c_id" jdbcType="VARCHAR" property="c_id"/>
+    <result column="c_name" jdbcType="VARCHAR" property="c_name"/>
+    <result column="c_count" property="c_count"/>
+  </resultMap>
+
+  <select id="findMyTeachCourse" parameterType="java.lang.String"
+            resultMap="courseItemShownOnTeacherEnd">
+
+    SELECT
+      course.c_id AS c_id,
+      course.c_name AS c_name,
+      c_count
+    FROM
+      course,
+      (SELECT c_id, COUNT(*) AS c_count FROM sc GROUP BY c_id) c_cnt
+    WHERE
+      course.`c_id` = c_cnt.`c_id` AND
+      course.`t_id` = #{t_id}
+  </select>
+
+  <select id="getStudentListByCourseId" parameterType="string"
+    resultType="student">
+    SELECT
+      *
+    FROM
+      student
+    WHERE
+        s_id IN (
+        SELECT sc.`s_id` FROM
+          sc
+        WHERE sc.`c_id` = #{c_id}
+      )
   </select>
 </mapper>
